Fix unversioned dep hashing on the initial run + README
When the cache breaker changed or on the initial run we wouldn't compute the hash of unversioned dependencies (e.g. local paths specified in package.json). Now include that on the initial run. This also fixes a small bug where the 2nd time you ran hyperinstall, it would think there were changes and run npm install in each package (which was a no-op but would still take time).

README was updated to show Hyperinstall's behavior so people can get a better sense of what it does.